Split builtin implementations
