+++ src/builtins_jobs.c
@@ -1,0 +1,54 @@
+/* Job control builtins */
+#define _GNU_SOURCE
+#include "builtins.h"
+#include "jobs.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <signal.h>
+
+int builtin_jobs(char **args) {
+    (void)args;
+    print_jobs();
+    return 1;
+}
+
+int builtin_fg(char **args) {
+    if (!args[1]) {
+        fprintf(stderr, "usage: fg ID\n");
+        return 1;
+    }
+    int id = atoi(args[1]);
+    wait_job(id);
+    return 1;
+}
+
+int builtin_bg(char **args) {
+    if (!args[1]) {
+        fprintf(stderr, "usage: bg ID\n");
+        return 1;
+    }
+    int id = atoi(args[1]);
+    bg_job(id);
+    return 1;
+}
+
+int builtin_kill(char **args) {
+    if (!args[1]) {
+        fprintf(stderr, "usage: kill [-SIGNAL] ID\n");
+        return 1;
+    }
+    int sig = SIGTERM;
+    int idx = 1;
+    if (args[1][0] == '-') {
+        sig = atoi(args[1] + 1);
+        idx++;
+    }
+    if (!args[idx]) {
+        fprintf(stderr, "usage: kill [-SIGNAL] ID\n");
+        return 1;
+    }
+    int id = atoi(args[idx]);
+    kill_job(id, sig);
+    return 1;
+}
+
